+++ v2023/src/stars/day_04.rs
@@ -12,6 +12,22 @@ pub fn scratchcard_points(input: String) -> u32 {
     parser::sum_iteration(input, point_counter)
 }
 
+pub fn scratchcard_replay(input: String) -> u32 {
+    let cards = input.trim().split('\n');
+    let mut replays = 0;
+    let mut pile = Vec::new();
+    pile.resize(cards.clone().count(), 1);
+    for (game, card) in cards.enumerate() {
+        let wins: usize = matching_numbers(card).try_into().unwrap();
+        let score = pile[game];
+        for value in pile.iter_mut().take(game + wins + 1).skip(game + 1) {
+            *value += score;
+        }
+        replays += score;
+    }
+    replays
+}
+
 fn matching_numbers(game: &str) -> u32 {
     let regex = Regex::new(r"(?<value>[0-9]+)").unwrap();
     let numbers: Vec<&str> = game.split([':', '|']).collect();
@@ -46,7 +62,19 @@ Card 4: 41 92 73 84 69 | 59 84 76 51 58  5 54 83
 Card 5: 87 83 26 28 32 | 88 30 70 12 93 22 82 36
 Card 6: 31 18 13 56 72 | 74 77 10 23 35 67 36 11";
     let value = scratchcard_points(input.to_string());
-    assert_eq!(value, 13)
+    assert_eq!(value, 13);
+}
+
+#[test]
+fn test_scratchcard_replay_example() {
+    let input = "Card 1: 41 48 83 86 17 | 83 86  6 31 17  9 48 53
+Card 2: 13 32 20 16 61 | 61 30 68 82 17 32 24 19
+Card 3:  1 21 53 59 44 | 69 82 63 72 16 21 14  1
+Card 4: 41 92 73 84 69 | 59 84 76 51 58  5 54 83
+Card 5: 87 83 26 28 32 | 88 30 70 12 93 22 82 36
+Card 6: 31 18 13 56 72 | 74 77 10 23 35 67 36 11";
+    let value = scratchcard_replay(input.to_string());
+    assert_eq!(value, 30);
 }
 
 #[test]
@@ -54,4 +82,11 @@ fn test_scratchcards_part_one() {
     let input = std::fs::read_to_string("./inputs/day04").unwrap();
     let actual = scratchcard_points(input);
     assert_eq!(actual, 19855);
+}
+
+#[test]
+fn test_scratchcards_part_two() {
+    let input = std::fs::read_to_string("./inputs/day04").unwrap();
+    let actual = scratchcard_replay(input);
+    assert_eq!(actual, 10378710);
 }